mysite/requirements.txt
Django==1.10.1
django-autocomplete-light==2.3.3
six==1.10.0
wheel==0.24.0
